server/requirements.txt
Flask
Flask_SQLAlchemy
Flask-Migrate
psycopg2-binary
flask_bcrypt
flask_jwt_extended
flask_cors
paho-mqtt
paho-mqtt<2.0.0